# finlogikk/back/src/main/java/com/finlogik/back/forum/entities/VoteComment.java
package com.finlogik.back.forum.entities;

import lombok.*;

import jakarta.persistence.*;

@Entity
@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@ToString
public class VoteComment {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private int id;
    @Enumerated(EnumType.STRING)
    private TypeVote typeVote;
    private String userId;
    private Long idComment;
}