# finlogikk/back/src/main/java/com/finlogik/back/forum/entities/VotePost.java
package com.finlogik.back.forum.entities;

import lombok.*;

import jakarta.persistence.*;

@Entity
@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@ToString
public class VotePost {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private int id;
    @Enumerated(EnumType.STRING)
    private TypeVote typeVote;
    private String userId;
    private Long idPost;
}