# finlogikk/back/src/main/java/com/finlogik/back/forum/entities/VoteTopic.java
package com.finlogik.back.forum.entities;

import lombok.*;

import jakarta.persistence.*;

@Entity
@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@ToString
public class VoteTopic {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private int id;
    @Enumerated(EnumType.STRING)
    private TypeVote typeVote;
    private String userId;
    private Long idTopic;

}