# finlogikk/front/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "yassine",
      "email": "atou26.ag@gmail.com"
    },
    "license": {
      "name": "Licence name",
      "url": "https://some-url.com"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8088/api/v1",
      "description": "Local ENV"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/voteTopic/VoteLike/{idTopic}/{userId}": {
      "post": {
        "tags": [
          "vote-topic-rest-controller"
        ],
        "operationId": "voteUserlike",
        "parameters": [
          {
            "name": "idTopic",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteTopic"
                }
              }
            }
          }
        }
      }
    },
    "/voteTopic/VoteDislike/{idTopic}/{userId}": {
      "post": {
        "tags": [
          "vote-topic-rest-controller"
        ],
        "operationId": "voteUserDislike",
        "parameters": [
          {
            "name": "idTopic",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteTopic"
                }
              }
            }
          }
        }
      }
    },
    "/votePost/VoteLike/{idPost}/{userId}": {
      "post": {
        "tags": [
          "vote-post-rest-controller"
        ],
        "operationId": "voteUserlike_1",
        "parameters": [
          {
            "name": "idPost",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VotePost"
                }
              }
            }
          }
        }
      }
    },
    "/votePost/VoteDislike/{idPost}/{userId}": {
      "post": {
        "tags": [
          "vote-post-rest-controller"
        ],
        "operationId": "voteUserDislike_1",
        "parameters": [
          {
            "name": "idPost",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VotePost"
                }
              }
            }
          }
        }
      }
    },
    "/voteComment/VoteLike/{idComment}/{userId}": {
      "post": {
        "tags": [
          "vote-comment-rest-controller"
        ],
        "operationId": "voteUserlike_2",
        "parameters": [
          {
            "name": "idComment",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteComment"
                }
              }
            }
          }
        }
      }
    },
    "/voteComment/VoteDislike/{idComment}/{use}": {
      "post": {
        "tags": [
          "vote-comment-rest-controller"
        ],
        "operationId": "voteUserDislike_2",
        "parameters": [
          {
            "name": "idComment",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteComment"
                }
              }
            }
          }
        }
      }
    },
...
    "/voteTopic/status/{topicId}/{userId}": {
      "get": {
        "tags": [
          "vote-topic-rest-controller"
        ],
        "operationId": "getUserVoteStatus",
        "parameters": [
          {
            "name": "topicId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/votePost/status/{PostId}/{userId}": {
      "get": {
        "tags": [
          "vote-post-rest-controller"
        ],
        "operationId": "getUserVoteStatus_1",
        "parameters": [
          {
            "name": "PostId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/voteComment/status/{CommentId}/{userId}": {
      "get": {
        "tags": [
          "vote-comment-rest-controller"
        ],
        "operationId": "getUserVoteStatus_2",
        "parameters": [
          {
            "name": "CommentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "VoteTopic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "typeVote": {
            "type": "string",
            "enum": [
              "LIKE",
              "DISLIKE",
              "NOVOTE"
            ]
          },
          "userId": {
            "type": "string"
          },
          "idTopic": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "VotePost": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "typeVote": {
            "type": "string",
            "enum": [
              "LIKE",
              "DISLIKE",
              "NOVOTE"
            ]
          },
          "userId": {
            "type": "string"
          },
          "idPost": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "VoteComment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "typeVote": {
            "type": "string",
            "enum": [
              "LIKE",
              "DISLIKE",
              "NOVOTE"
            ]
          },
          "userId": {
            "type": "string"
          },
          "idComment": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "oauth2",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "flows": {
          "clientCredentials": {
            "authorizationUrl": "http://localhost:9090/realms/finlogik/protocol/openid-connect/auth",
            "scopes": {}
          }
        }
      }
    }
  }
}